[DCOS-56081] Make "service configuration JSON"-standardizing script more flexible. (#3122)
This allows for the script to also standardize project specific sections. Right
now it only standardizes the order of the "service" section.

For example, I'm using it to standardize the order of the "elasticsearch"
section for the Elastic service. By providing a configuration file that looks
like the following:

`sdk-tools.json`:
```
{
  "standardize_config_json": {
    "sections": {
      "elasticsearch": {
        "head": [
          "custom_elasticsearch_yml",
          "health_user",
          "health_user_password",
          "plugins",
          "plugin_http_proxy_host",
          "plugin_http_proxy_port",
          "plugin_https_proxy_host",
          "plugin_https_proxy_port",
          "xpack_http_proxy_host",
          "xpack_http_proxy_port",
          "xpack_graph_enabled",
          "xpack_ml_enabled",
          "xpack_monitoring_collection_enabled",
          "xpack_monitoring_enabled",
          "xpack_security_audit_enabled",
          "xpack_security_enabled",
          "xpack_sql_enabled",
          "xpack_watcher_enabled"
        ],
        "tail": ["http_content_type_required", "xpack_enabled"]
      }
    }
  }
}
```

and calling the standardization script like:

```
./dcos-commons/tools/standardize_config_json.py frameworks/elastic/universe/config.json sdk-tools.json
```

the script will make sure that:
1. the first elements in the "elasticsearch" section are the ones in the "head" section
2. the last elements in the "elasticsearch" section are the ones in the "tail" section
3. the remaining elements (in the middle) will be sorted alphabetically

The "head" section is usually reserved for important settings, and the "tail"
section for deprecated settings.
